+++ dotnet/src/Connectors/Connectors.HuggingFace/Client/TextEmbeddingResponse.cs
@@ -3,11 +3,11 @@
 using System;
 using System.Collections.Generic;
 
-namespace Microsoft.SemanticKernel.Connectors.HuggingFace;
+namespace Microsoft.SemanticKernel.Connectors.HuggingFace.Client;
 
 /// <summary>
 /// Represents the response from the Hugging Face text embedding API.
 /// </summary>
-public sealed class TextEmbeddingResponse : List<List<List<ReadOnlyMemory<float>>>>
+internal sealed class TextEmbeddingResponse : List<List<List<ReadOnlyMemory<float>>>>
 {
 }
