+++ dotnet/src/Connectors/Connectors.HuggingFace/Client/TextGenerationResponse.cs
@@ -2,13 +2,13 @@
 
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
-using static Microsoft.SemanticKernel.Connectors.HuggingFace.Core.TextGenerationResponse;
+using static Microsoft.SemanticKernel.Connectors.HuggingFace.Client.TextGenerationResponse;
 
 #pragma warning disable CA1812 // Avoid uninstantiated internal classes
 
-namespace Microsoft.SemanticKernel.Connectors.HuggingFace.ImageToText;
+namespace Microsoft.SemanticKernel.Connectors.HuggingFace.Client;
 
-internal sealed class ImageToTextGenerationResponse : List<GeneratedTextItem>
+internal sealed class TextGenerationResponse : List<GeneratedTextItem>
 {
     internal sealed class GeneratedTextItem
     {
