+++ dotnet/src/Connectors/Connectors.HuggingFace/HuggingFacePromptExecutionSettings.cs
