+++ dotnet/src/Connectors/Connectors.HuggingFace/Services/HuggingFaceImageToTextService.cs
@@ -1,0 +1,58 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+using System;
+using System.Collections.Generic;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+using Microsoft.SemanticKernel.Connectors.HuggingFace.Client;
+using Microsoft.SemanticKernel.Http;
+using Microsoft.SemanticKernel.ImageToText;
+using Microsoft.SemanticKernel.Services;
+
+namespace Microsoft.SemanticKernel.Connectors.HuggingFace;
+
+/// <summary>
+/// HuggingFace image to text service
+/// </summary>
+public sealed class HuggingFaceImageToTextService : IImageToTextService
+{
+    private readonly Dictionary<string, object?> _attributesInternal = new();
+    private readonly HuggingFaceClient _client;
+
+    /// <inheritdoc />
+    public IReadOnlyDictionary<string, object?> Attributes => this._attributesInternal;
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="HuggingFaceImageToTextService"/> class.
+    /// </summary>
+    /// <param name="model">The HuggingFace model for image-to-text conversion.</param>
+    /// <param name="endpoint">The endpoint uri including the port where HuggingFace server is hosted</param>
+    /// <param name="apiKey">Optional API key for accessing the HuggingFace service.</param>
+    /// <param name="httpClient">Optional HTTP client to be used for communication with the HuggingFace API.</param>
+    /// <param name="loggerFactory">Optional logger factory to be used for logging.</param>
+    public HuggingFaceImageToTextService(
+        string model,
+        Uri? endpoint = null,
+        string? apiKey = null,
+        HttpClient? httpClient = null,
+        ILoggerFactory? loggerFactory = null)
+    {
+        Verify.NotNullOrWhiteSpace(model);
+
+        this._client = new HuggingFaceClient(
+            modelId: model,
+            endpoint: endpoint ?? httpClient?.BaseAddress,
+            apiKey: apiKey,
+            httpClient: HttpClientProvider.GetHttpClient(httpClient),
+            logger: loggerFactory?.CreateLogger(this.GetType())
+        );
+
+        this._attributesInternal.Add(AIServiceExtensions.ModelIdKey, model);
+    }
+
+    /// <inheritdoc />
+    public Task<IReadOnlyList<TextContent>> GetTextContentsAsync(ImageContent content, PromptExecutionSettings? executionSettings = null, Kernel? kernel = null, CancellationToken cancellationToken = default)
+        => this._client.GenerateTextFromImageAsync(content, executionSettings, kernel, cancellationToken);
+}
