+++ dotnet/src/Connectors/Connectors.HuggingFace/Services/HuggingFaceTextEmbeddingGenerationService.cs
@@ -6,33 +6,33 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using Microsoft.SemanticKernel.Connectors.HuggingFace.Core;
+using Microsoft.SemanticKernel.Connectors.HuggingFace.Client;
+using Microsoft.SemanticKernel.Embeddings;
 using Microsoft.SemanticKernel.Http;
-using Microsoft.SemanticKernel.ImageToText;
 using Microsoft.SemanticKernel.Services;
 
 namespace Microsoft.SemanticKernel.Connectors.HuggingFace;
 
 /// <summary>
-/// HuggingFace image to text service
+/// HuggingFace embedding generation service.
 /// </summary>
-public sealed class HuggingFaceImageToTextService : IImageToTextService
+public sealed class HuggingFaceTextEmbeddingGenerationService : ITextEmbeddingGenerationService
 {
-    private readonly Dictionary<string, object?> _attributesInternal = new();
-    private readonly HuggingFaceClient _client;
+    private Dictionary<string, object?> AttributesInternal { get; } = new();
+    private HuggingFaceClient Client { get; }
 
     /// <inheritdoc />
-    public IReadOnlyDictionary<string, object?> Attributes => this._attributesInternal;
+    public IReadOnlyDictionary<string, object?> Attributes => this.AttributesInternal;
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="HuggingFaceImageToTextService"/> class.
+    /// Initializes a new instance of the <see cref="HuggingFaceTextEmbeddingGenerationService"/> class.
     /// </summary>
-    /// <param name="model">The HuggingFace model for image-to-text conversion.</param>
+    /// <param name="model">The HuggingFace model for the text generation service.</param>
     /// <param name="endpoint">The endpoint uri including the port where HuggingFace server is hosted</param>
     /// <param name="apiKey">Optional API key for accessing the HuggingFace service.</param>
     /// <param name="httpClient">Optional HTTP client to be used for communication with the HuggingFace API.</param>
     /// <param name="loggerFactory">Optional logger factory to be used for logging.</param>
-    public HuggingFaceImageToTextService(
+    public HuggingFaceTextEmbeddingGenerationService(
         string model,
         Uri? endpoint = null,
         string? apiKey = null,
@@ -41,20 +41,18 @@ public sealed class HuggingFaceImageToTextService : IImageToTextService
     {
         Verify.NotNullOrWhiteSpace(model);
 
-        this._client = new HuggingFaceClient(
-            modelId: model,
+        this.Client = new HuggingFaceClient(
+        modelId: model,
             endpoint: endpoint ?? httpClient?.BaseAddress,
             apiKey: apiKey,
-#pragma warning disable CA2000 // Dispose objects before losing scope
             httpClient: HttpClientProvider.GetHttpClient(httpClient),
-#pragma warning restore CA2000 // Dispose objects before losing scope
             logger: loggerFactory?.CreateLogger(this.GetType())
         );
 
-        this._attributesInternal.Add(AIServiceExtensions.ModelIdKey, model);
+        this.AttributesInternal.Add(AIServiceExtensions.ModelIdKey, model);
     }
 
-    /// <inheritdoc />
-    public Task<IReadOnlyList<TextContent>> GetTextContentsAsync(ImageContent content, PromptExecutionSettings? executionSettings = null, Kernel? kernel = null, CancellationToken cancellationToken = default)
-        => this._client.GenerateTextFromImageAsync(content, executionSettings, kernel, cancellationToken);
+    /// <inheritdoc/>
+    public Task<IList<ReadOnlyMemory<float>>> GenerateEmbeddingsAsync(IList<string> data, Kernel? kernel = null, CancellationToken cancellationToken = default)
+        => this.Client.GenerateEmbeddingsAsync(data, kernel, cancellationToken);
 }
