+++ dotnet/src/Connectors/Connectors.HuggingFace/TextGeneration/TextGenerationStreamMetadata.cs
@@ -3,7 +3,7 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Runtime.CompilerServices;
-using Microsoft.SemanticKernel.Connectors.HuggingFace.Core;
+using Microsoft.SemanticKernel.Connectors.HuggingFace.Client;
 
 namespace Microsoft.SemanticKernel.Connectors.HuggingFace.TextGeneration;
 
@@ -51,7 +51,7 @@ public sealed class TextGenerationStreamMetadata : ReadOnlyDictionary<string, ob
     }
 
     /// <summary>
-    /// Unique token idenfitier for the model
+    /// Unique token identifier for the model
     /// </summary>
     public int? TokenId
     {
